project-01-ai-agent/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.0",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
